Refactor routes (#20)
* Refactor authentication to write to database in middlewares

It's a cleaner implementation to handle writing to the
database in the encapsulated middlewares. This is also a step
towards using express-router and controllers

* Rename auth_errorHandler

* Add enter router for login/signup

For better modularity, routes using similar paths
and middlewares should be refactored to use express
routers and controllers. This the first step

* Add routes to change user settings

User should be able to change their email, password,
and delete their account

* Change database references from 'local' to 'db'

* Add routes for getting user posts

App should be able to retrieve user's saved posts and the
posts the user made

* Add remaining post routes and small implementation changes

Add router/controller for posts routes. Changed database to
load data automatically when it's required with 'require'.
Left suggestions on delete-post and get-post.

* Remove uneccessary 'async' users in controllers

* Remove uneeded routes.js file

All routes have been written in routers and controllers

* Refactor index.js

Remove now uneeded code and rename to server.js

* Implemented all routes except change password, change email, delete account, and search.

 * Database data is now only loaded once on server start.
 * added testing framework for authorized requests
 * altered delete post endpoint so that clients can only delete posts that they created
 * Fixed doc errors

* Fixed leftover query -> req.params

Co-authored-by: EinarBalan <einar.balan@gmail.com>
Co-authored-by: Einar Balan <40647794+EinarBalan@users.noreply.github.com>